Eliminates some dependencies...
Between installed files and the directories they must be in.
When you install a file in a directory normally this also modifies
the directory's timestamp, so some rules are always triggered and
this makes installation of some files to occur always, while others
for some reason are not.

Anyway, if make exits on some file not being able to install, you
have also targets defined to create the installation directories
needed.

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>